Centralizes BootstrapVariant type definition
Moves the BootstrapVariant type definition to a common index file and imports it in the files that previously had duplicate definitions.

Eliminates code duplication and ensures consistency across components by having a single source of truth for Bootstrap variant types.

Adds missing "outline-link" variant to the centralized definition.
